Fix weird cursor bug
Bug: When opening the app, the cursor was often just at a random location

Cause: Input state is only being updated when app_state is not loading. When the loading scene transitions to the home scene, app_state is changed to idle AFTER the chance to update the input state. This means that, for one tick, the input state is just garbage

Fix: Just update the input state regardless of app_state??? Why was this not already happening